+++ src/machine_learning_test.rs
@@ -6,3 +6,4 @@ mod meanshift_test;
 mod dbscan_test;
 mod decision_tree_test;
 mod isolation_tree_test;
+mod principal_component_analysis_test;